% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dna_data.R
\name{add_any_amp}
\alias{add_any_amp}
\title{Adds Any Amp rows, which are rows with AMP, HighAMP, and FocalHighAMP
in the alt column}
\usage{
add_any_amp(.data)
}
\arguments{
\item{.data}{A data frame with an 'alt' column}
}
\value{
A data frame with 'Any Amp' rows
}
\description{
Adds Any Amp rows, which are rows with AMP, HighAMP, and FocalHighAMP
in the alt column
}
\examples{
add_any_amp(get_mbcp_dna_alts())
}
