% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_data.R
\name{get_clinical_data}
\alias{get_clinical_data}
\title{Get clinical data with selected variables}
\usage{
get_clinical_data()
}
\value{
A data frame with selected variables
}
\description{
Get clinical data with selected variables
}
\examples{
get_clinical_data()
}
