% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_data.R
\name{get_clinical_data_all_variables}
\alias{get_clinical_data_all_variables}
\title{Get clinical data with all the variables available}
\usage{
get_clinical_data_all_variables()
}
\value{
A data frame with all the variables available
}
\description{
Get clinical data with all the variables available
}
\examples{
get_clinical_data_all_variables()
}
