% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_data.R
\name{get_clinical_data_first_samples}
\alias{get_clinical_data_first_samples}
\title{Get clinical data for first samples per patient}
\usage{
get_clinical_data_first_samples(sample_type = "any")
}
\arguments{
\item{sample_type}{Character value to specify the type of samples to include. Options are:
\itemize{
\item{any:}{gets the first timepoint sample per patient}
\item{wes:}{gets the minimum timepoint sample per patient, only for WES samples}
\item{rna:}{gets the minimum timepoint sample per patient, only for RNA samples}
\item{wes_rna:}{gets the minimum timepoint sample per patient, for samples with WES and RNA}
}}
}
\value{
A data frame with sample level clinical information
}
\description{
Get clinical data for first samples per patient
}
\examples{
get_clinical_data_first_samples(sample_type = "any")
get_clinical_data_first_samples(sample_type = "wes_rna")
}
