% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_data.R
\name{get_clinical_data_multisamples}
\alias{get_clinical_data_multisamples}
\title{Get clinical data for patients with multiple samples in reference/sample format}
\usage{
get_clinical_data_multisamples()
}
\value{
A data frame with one entry per reference and comparison sample pair.
Reference samples are first samples with WES in patients where n_bypatient > 1 and they
are paired with the remaining WES samples from the same patient.
}
\description{
Get clinical data for patients with multiple samples in reference/sample format
}
\examples{
get_clinical_data_multisamples()
}
