% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_data.R
\name{get_clinical_data_wes}
\alias{get_clinical_data_wes}
\title{Get clinical data for WES samples}
\usage{
get_clinical_data_wes()
}
\value{
A data frame with sample level clinical information
}
\description{
Get clinical data for WES samples
}
\examples{
get_clinical_data_wes()
}
