% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mbcp_colors.R
\name{get_mbcp_colors}
\alias{get_mbcp_colors}
\title{Get color palettes for MBC project data}
\usage{
get_mbcp_colors(set)
}
\arguments{
\item{set}{Character string specifying the color palette to return. Options are: "receptor", "pam50",
"dna_alts", "mutation", "histology", "treatment", "metastasis", "categories", "germline_alts", "dna_signatures"}
}
\value{
A named vector of colors
}
\description{
Get color palettes for MBC project data
}
\examples{
get_mbcp_colors("receptor")
}
