% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dna_data.R
\name{get_mbcp_dna_alts}
\alias{get_mbcp_dna_alts}
\title{Get DNA alts found in MBC project}
\usage{
get_mbcp_dna_alts(samples = NULL, genes = NULL, min_ccf = 0)
}
\arguments{
\item{samples}{Character vector of samples to filter variations by sample}

\item{genes}{Character vector of hugo symbols to filter variations by gene}

\item{min_ccf}{Minimum ccf}
}
\value{
A data frame with DNA alts. Variant types: MUT (snv) y CNV
}
\description{
Get DNA alts found in MBC project
}
\examples{
get_mbcp_dna_alts()
}
