% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dna_data.R
\name{get_mbcp_dna_alts_summary}
\alias{get_mbcp_dna_alts_summary}
\title{Get dna alterations (CNVs and mutations) in the MBC project.}
\usage{
get_mbcp_dna_alts_summary(
  samples = NULL,
  genes = NULL,
  min_ccf = NULL,
  variant_type = "both"
)
}
\arguments{
\item{samples}{Character vector of samples to filter variations by sample}

\item{genes}{Character vector of hugo symbols to filter variations by gene}

\item{min_ccf}{Minimum cancer cell fraction}

\item{variant_type}{Used to specify the type of variant alterations to add in the summary. Must
be one of: "both", "CNV", "MUT"

One entry per sample and gene for mutations with Multiple_hit, if multiple entries
One entry per sample and gene for amplifications,}
}
\description{
Get dna alterations (CNVs and mutations) in the MBC project.
}
\examples{
get_mbcp_dna_alts_summary(genes = c("ESR1", "TP53"), min_ccf = 0.75)
}
