% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enrichment_data.R
\name{get_mbcp_enrichment_class}
\alias{get_mbcp_enrichment_class}
\title{Get MBCproject enrichment classification}
\usage{
get_mbcp_enrichment_class(samples = NULL, signatures = NULL)
}
\arguments{
\item{samples}{Character vector of samples to filter columns}

\item{signatures}{Character vector of signature ids to filter rows}
}
\value{
A character matrix with classification of normalized enrichment scores for hallmark and selected sets in the MBCproject
}
\description{
Get MBCproject enrichment classification
}
\examples{
get_mbcp_enrichment_class()
}
