% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature_data.R
\name{get_mbcp_expression_signature_class}
\alias{get_mbcp_expression_signature_class}
\title{Get MBCproject enrichment classification}
\usage{
get_mbcp_expression_signature_class(signatures = NULL, samples = NULL)
}
\arguments{
\item{signatures}{Character vector of signature ids to filter rows}

\item{samples}{Character vector of samples to filter columns}
}
\value{
A character matrix with classification of normalized enrichment scores for hallmark signatures and other selected sets.
Categories are: Lower Decile, Lower Quartile, Upper Quartile, and Upper Decile. \code{NA} values for entries that do not fall into these categories.
}
\description{
Get MBCproject enrichment classification
}
\examples{
get_mbcp_expression_signature_class()
}
