% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature_data.R
\name{get_mbcp_expression_signatures}
\alias{get_mbcp_expression_signatures}
\title{Get list of signatures and their associated genes studied in the MBCProject}
\usage{
get_mbcp_expression_signatures()
}
\value{
A data frame with signature, gene, and category columns for genes functionally
associated with their signatures.
}
\description{
Get list of signatures and their associated genes studied in the MBCProject
}
\examples{
get_mbcp_expression_signatures()
}
