% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dna_data.R
\name{get_mbcp_germline_hits}
\alias{get_mbcp_germline_hits}
\title{Get germline hits dataset}
\usage{
get_mbcp_germline_hits()
}
\value{
A data frame with entries from the germline hits found in the MBCProject
}
\description{
Get germline hits dataset
}
\examples{
get_mbcp_germline_hits()
}
