% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mbcp_labels.R
\name{get_mbcp_labels}
\alias{get_mbcp_labels}
\title{Get label vectors for MBC project data}
\usage{
get_mbcp_labels(set)
}
\arguments{
\item{set}{Character string specifying the labels to return. Options are: "receptor", "pam50",
"dna_alts", "mutation", "histology", "treatment", "metastasis", "expression"}
}
\value{
A named character vector
}
\description{
Get label vectors for MBC project data
}
\examples{
get_mbcp_labels("receptor")
}
