% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_data.R
\name{get_mbcp_log2p1_tpms_class}
\alias{get_mbcp_log2p1_tpms_class}
\title{Get gene expression classification}
\usage{
get_mbcp_log2p1_tpms_class(genes = NULL, samples = NULL)
}
\arguments{
\item{genes}{Character vector of hugo symbols to filter rows}

\item{samples}{Character vector of samples to filter columns}
}
\value{
A character matrix with classification of log2 tpms.
Categories are: Lower Decile, Lower Quartile, Upper Quartile, and Upper Decile. \code{NA} values for entries that do not fall into these categories.
}
\description{
Get gene expression classification
}
\examples{
get_mbcp_log2p1_tpms_class()
}
