% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dna_data.R
\name{get_mbcp_mutations}
\alias{get_mbcp_mutations}
\title{Get mutations in MBC project}
\usage{
get_mbcp_mutations(samples = NULL, genes = NULL, min_ccf = NULL)
}
\arguments{
\item{samples}{Character vector of samples to filter variations by sample}

\item{genes}{Character vector of hugo symbols to filter variations by gene}

\item{min_ccf}{Minimum cancer cell fraction}
}
\value{
A data frame with mutations in a MAF format
}
\description{
Get mutations in MBC project
}
\examples{
get_mbcp_mutations()
}
