% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dna_data.R
\name{get_mbcp_snvs}
\alias{get_mbcp_snvs}
\title{Get SNVs found in MBC project}
\usage{
get_mbcp_snvs(samples = NULL, genes = NULL, min_ccf = 0)
}
\arguments{
\item{samples}{Character vector of samples to filter variations by sample}

\item{genes}{Character vector of hugo symbols to filter variations by gene}

\item{min_ccf}{Minimum ccf}
}
\value{
A data frame with SNVs
}
\description{
Get SNVs found in MBC project
}
\examples{
get_mbcp_snvs()
}
