% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_data.R
\name{get_mbcp_tpms_uq}
\alias{get_mbcp_tpms_uq}
\title{Get upper quartile normalization TPMs}
\usage{
get_mbcp_tpms_uq(samples = NULL, genes = NULL)
}
\arguments{
\item{samples}{Character vector of samples to filter columns}

\item{genes}{Character vector of hugo symbols to filter rows}
}
\value{
A matrix of UQ TPMs with genes as rows and samples as columns with sample_alias as column names
}
\description{
Get upper quartile normalization TPMs
}
\examples{
get_mbcp_tpms_uq()
}
