% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbcp_data.R
\docType{data}
\name{mbcp_clinical_data}
\alias{mbcp_clinical_data}
\title{MBCproject clinical data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 421 rows and 212 columns.
}
\usage{
mbcp_clinical_data
}
\description{
Data frame with clinical data for each RNA or Wes sample in the MBCproject
}
\keyword{datasets}
