% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbcp_data.R
\docType{data}
\name{mbcp_cnvs}
\alias{mbcp_cnvs}
\title{MBCproject copy number variations
Data frame in gene level seg file format, plus columns added by ABSOLUTE and
ONCOKB annotator}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3204743 rows and 48 columns.
}
\usage{
mbcp_cnvs
}
\description{
MBCproject copy number variations
Data frame in gene level seg file format, plus columns added by ABSOLUTE and
ONCOKB annotator
}
\keyword{datasets}
