% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbcp_data.R
\docType{data}
\name{mbcp_dna_alts}
\alias{mbcp_dna_alts}
\title{MBCproject DNA alts}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1295673 rows and 7 columns.
}
\usage{
mbcp_dna_alts
}
\description{
Data frame with DNA alterations found in MBCproject
It includes SNVs and CNVs from MAF and SEG files
}
\keyword{datasets}
