% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbcp_data.R
\docType{data}
\name{mbcp_enrichment_class}
\alias{mbcp_enrichment_class}
\title{MBCproject enrichment classification}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 231 rows and 200 columns.
}
\usage{
mbcp_enrichment_class
}
\description{
Classification of normalized enrichment scores for hallmark and selected sets in the MBCproject
obtained using a method for cohort balancing according to receptor status
Categories: Lower Decile, Lower Quartile, Upper Quartile, Upper Decile.
}
\keyword{datasets}
