% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbcp_data.R
\docType{data}
\name{mbcp_enrichment_scores}
\alias{mbcp_enrichment_scores}
\title{MBCproject enrichment scores}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 229 rows and 200 columns.
}
\usage{
mbcp_enrichment_scores
}
\description{
Numeric matrix with normalized enrichment scores for hallmark and selected sets in the MBCproject
Rows are gene sets and columns are sample aliases. Scores were obtained using the
GSEA implementation in the fgsea package
}
\keyword{datasets}
