% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbcp_data.R
\docType{data}
\name{mbcp_germline_hits}
\alias{mbcp_germline_hits}
\title{MBCproject germline hits}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 8 columns.
}
\usage{
mbcp_germline_hits
}
\description{
Must remove this for public release
}
\keyword{datasets}
