% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbcp_data.R
\docType{data}
\name{mbcp_log2_tpms}
\alias{mbcp_log2_tpms}
\title{MBCproject log2 TPMs}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 12195 rows and 200 columns.
}
\usage{
mbcp_log2_tpms
}
\description{
Numeric matrix with log2 TPMs for each gene in the MBCproject.
Rows are genes identified by hugo symbol and columns are sample aliases
}
\keyword{datasets}
