% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbcp_data.R
\docType{data}
\name{mbcp_log2p1_tpms_class}
\alias{mbcp_log2p1_tpms_class}
\title{MBCproject log2 tpms classification}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 12071 rows and 200 columns.
}
\usage{
mbcp_log2p1_tpms_class
}
\description{
Classification of log2 tpms values of gene expression
obtained using a method for cohort balancing according to receptor status
Categories: Lower Decile, Lower Quartile, Upper Quartile, Upper Decile.
\code{NA} values for entries that do not fall into these categories
}
\keyword{datasets}
