% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbcp_data.R
\docType{data}
\name{mbcp_mutations}
\alias{mbcp_mutations}
\title{MBCproject mutations}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 29950 rows and 196 columns.
}
\usage{
mbcp_mutations
}
\description{
Data frame in MAF format with mutations found in MBCproject.
It includes mandatory fields in a MAF file, plus columns added by ABSOLUTE
and ONCOKB annotator
}
\keyword{datasets}
