% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_data.R
\name{to_longer}
\alias{to_longer}
\title{Get matrix with mbcp data in long format}
\usage{
to_longer(mbcp_matrix, id_column)
}
\arguments{
\item{mbcp_matrix}{Numeric matrix with TPMs or Signature data}

\item{id_column}{Character value to identify rownames in matrix. Options are: "hugo_symbol" and "signature_id"}
}
\value{
A dataframe with values in long format. The number of rows equals \var{n}x\var{m}, the dimensions
of \code{mbcp_matrix}, with three columns:
\itemize{
\item{Value of \code{id_column}:} {Contains rownames in the \code{mbcp_matrix}}
\item{\var{sample_alias}:} {Contains colnames in the \code{mbcp_matrix}}
\item{\var{value}:} {The [\code{id_column}, \var{sample_alias}] value in the \code{mbcp_matrix}}
}
}
\description{
Get matrix with mbcp data in long format
}
\examples{
to_longer(get_mbcp_tpms_uq(), "hugo_symbol")
}
