% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dna_data.R
\name{get_mbcp_cnvs}
\alias{get_mbcp_cnvs}
\title{Get CNV in MBC project}
\usage{
get_mbcp_cnvs(samples = NULL, genes = NULL)
}
\arguments{
\item{samples}{Character vector of samples to filter variations by sample}

\item{genes}{Character vector of hugo symbols to filter variations by gene}
}
\description{
Get CNV in MBC project
}
\examples{
get_mbcp_cnvs()
}
