% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dna_data.R
\name{get_mbcp_dna_alts_summary}
\alias{get_mbcp_dna_alts_summary}
\title{Get dna alterations (CNVs and mutations) in the MBC project.}
\usage{
get_mbcp_dna_alts_summary(samples = NULL, genes = NULL, min_ccf = NULL)
}
\arguments{
\item{samples}{Character vector of samples to filter variations by sample}

\item{genes}{Character vector of hugo symbols to filter variations by gene}

\item{min_ccf}{Minimum cancer cell fraction}
}
\description{
Get dna alterations (CNVs and mutations) in the MBC project.
}
\examples{
get_mbcp_dna_alts_summary(min_ccf = 0.75)
}
