% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enrichment_data.R
\name{get_mbcp_enrichment_scores}
\alias{get_mbcp_enrichment_scores}
\title{Get MBCproject enrichment scores}
\usage{
get_mbcp_enrichment_scores(signatures = NULL, samples = NULL)
}
\arguments{
\item{signatures}{Character vector of signature ids to filter rows}

\item{samples}{Character vector of samples to filter columns}
}
\value{
A numeric matrix with normalized enrichment scores for hallmark and selected sets in the MBCproject
}
\description{
Get MBCproject enrichment scores
}
\examples{
get_mbcp_enrichment_scores()
}
