% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_data.R
\name{get_mbcp_tpms}
\alias{get_mbcp_tpms}
\title{Get TPMs}
\usage{
get_mbcp_tpms(genes = NULL, samples = NULL)
}
\arguments{
\item{genes}{Character vector of hugo symbols to filter rows}

\item{samples}{Character vector of samples to filter columns}
}
\value{
A matrix of TPMs with genes as rows and samples as columns with sample_alias as column names
}
\description{
Get TPMs
}
\examples{
get_mbcp_tpms()
}
